requirements.txt
pillow
requests
ollama
ipykernel
googletrans
